Move tools to the common project and restore tests.
